+++ kotlin-client/src/main/kotlin/restwars/client/RestWarsClient.kt
@@ -3,98 +3,22 @@ package restwars.client
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
-import feign.Feign
-import feign.Headers
-import feign.Param
-import feign.RequestLine
-import feign.auth.BasicAuthRequestInterceptor
-import feign.jackson.JacksonDecoder
-import feign.jackson.JacksonEncoder
-import feign.slf4j.Slf4jLogger
-import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage
-import org.eclipse.jetty.websocket.api.annotations.WebSocket
-import org.eclipse.jetty.websocket.client.ClientUpgradeRequest
-import org.eclipse.jetty.websocket.client.WebSocketClient
 import restwars.rest.api.*
-import java.net.URI
-import java.util.concurrent.ConcurrentSkipListSet
-
-interface RoundCallback {
-    fun callback(response: RoundResponse)
-}
 
 open class RestWarsClient(val hostname: String, val port: Int) {
     protected val httpBaseUrl = "http://$hostname:$port/"
-    private val websocketBaseUrl = "ws://$hostname:$port/"
+    private val webSocketBaseUrl = "ws://$hostname:$port/"
 
-    private val mapper = ObjectMapper()
+    protected val mapper = ObjectMapper()
             .registerModule(KotlinModule())
             .configure(DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES, false)
 
-    private val client: Restwars = feignBuilder()
-            .target(Restwars::class.java, httpBaseUrl)
+    private val client: Restwars = Restwars.create(httpBaseUrl, mapper)
 
-    private val roundCallbacks: MutableSet<RoundCallback> = ConcurrentSkipListSet()
+    private val roundWebSocket = WebSocket(mapper, RoundResponse::class.java, webSocketBaseUrl + "v1/round/websocket")
+    private val tournamentWebSocket = WebSocket(mapper, SuccessResponse::class.java, webSocketBaseUrl + "v1/tournament/websocket")
 
-    private var websocketClient: WebSocketClient? = null
-    private val websocketClientLock = Object()
-
-    @Headers("Content-Type: application/json", "Accept: application/json")
-    interface Restwars {
-        @RequestLine("POST /v1/player")
-        fun createPlayer(request: CreatePlayerRequest)
-
-        @RequestLine("GET /v1/planet")
-        fun listPlanets(): PlanetsResponse
-
-        @RequestLine("GET /v1/player/fight")
-        fun listFights(): FightsResponse
-
-        @RequestLine("GET /v1/planet/{location}/fight")
-        fun listFightsWithPlanet(@Param("location") location: String): FightsResponse
-
-        @RequestLine("POST /v1/planet/{location}/telescope/scan")
-        fun telescopeScan(@Param("location") location: String): ScanResponse
-
-        @RequestLine("GET /v1/restwars")
-        fun applicationInformation(): ApplicationInformationResponse
-
-        @RequestLine("GET /v1/configuration")
-        fun restwarsConfiguration(): ConfigResponse
-
-        @RequestLine("GET /v1/round")
-        fun roundInformation(): RoundWithRoundTimeResponse
-
-        @RequestLine("GET /v1/planet/{location}/building")
-        fun listBuildings(@Param("location") location: String): BuildingsResponse
-
-        @RequestLine("POST /v1/planet/{location}/building")
-        fun constructBuilding(@Param("location") location: String, request: BuildBuildingRequest): ConstructionSiteResponse
-
-        @RequestLine("GET /v1/planet/{location}/construction-site")
-        fun listConstructionSites(@Param("location") location: String): ConstructionSitesResponse
-
-        @RequestLine("GET /v1/planet/{location}/hangar")
-        fun listShipsInHangar(@Param("location") location: String): ShipsResponse
-
-        @RequestLine("POST /v1/planet/{location}/hangar")
-        fun constructShip(@Param("location") location: String, request: BuildShipRequest): ShipInConstructionResponse
-
-        @RequestLine("GET /v1/planet/{location}/shipyard")
-        fun listShipsInConstruction(@Param("location") location: String): ShipsInConstructionResponse
-
-        @RequestLine("POST /v1/planet/{location}/flight")
-        fun startFlight(@Param("location") location: String, request: CreateFlightRequest): FlightResponse
-
-        @RequestLine("GET /v1/flight")
-        fun listFlights(): FlightsResponse
-
-        @RequestLine("GET /v1/flight/to/{location}")
-        fun listFlightsWithDestination(@Param("location") location: String): FlightsResponse
-
-        @RequestLine("GET /v1/flight/from/{location}")
-        fun listFlightsWithStart(@Param("location") location: String): FlightsResponse
-    }
+    fun getPlayer(): PlayerResponse = client.getPlayer()
 
     fun createPlayer(username: String, password: String) {
         client.createPlayer(CreatePlayerRequest(username, password))
@@ -106,6 +30,16 @@ open class RestWarsClient(val hostname: String, val port: Int) {
 
     fun roundInformation(): RoundWithRoundTimeResponse = client.roundInformation()
 
+    fun waitForNextRound(): RoundResponse = client.waitForNextRound()
+
+    fun listShipMetadata(): ShipsMetadataResponse = client.listShipMetadata()
+
+    fun listBuildingMetadat(): BuildingsMetadataResponse = client.listBuildingMetadata()
+
+    fun waitForTournamentStart(): SuccessResponse = client.waitForTournamentStart()
+
+    fun listsPoints(): PointsResponse = client.listsPoints()
+
     fun withCredentials(username: String, password: String): AuthenticatingRestWarsClient {
         return AuthenticatingRestWarsClient(hostname, port, username, password)
     }
@@ -113,73 +47,61 @@ open class RestWarsClient(val hostname: String, val port: Int) {
     /**
      * Adds a [callback] which is called on the start of a new round.
      */
-    fun addRoundCallback(callback: RoundCallback) {
-        roundCallbacks.add(callback)
+    fun addRoundCallback(callback: WebSocketCallback<RoundResponse>) {
+        roundWebSocket.addCallback(callback)
+    }
 
-        synchronized(websocketClientLock) {
-            if (websocketClient == null) {
-                startWebsocketListener()
+    /**
+     * Adds a [callback] which is called on the start of a new round.
+     */
+    fun addRoundCallback(callback: (RoundResponse) -> Unit) {
+        addRoundCallback(object : WebSocketCallback<RoundResponse> {
+            override fun call(response: RoundResponse) {
+                callback(response)
             }
-        }
+        })
     }
 
     /**
      * Removes a round [callback].
      */
-    fun removeRoundCallback(callback: RoundCallback) {
-        roundCallbacks.remove(callback)
+    fun removeRoundCallback(callback: WebSocketCallback<RoundResponse>) {
+        roundWebSocket.removeCallback(callback)
+    }
 
-        if (roundCallbacks.isEmpty()) {
-            synchronized(websocketClientLock) {
-                stopWebsocketListener()
+    /**
+     * Adds a [callback] which is called on the start of the tournament.
+     */
+    fun addTournamentCallback(callback: WebSocketCallback<SuccessResponse>) {
+        tournamentWebSocket.addCallback(callback)
+    }
+
+    /**
+     * Adds a [callback] which is called on the start of the tournament.
+     */
+    fun addTournamentCallback(callback: (SuccessResponse) -> Unit) {
+        addTournamentCallback(object : WebSocketCallback<SuccessResponse> {
+            override fun call(response: SuccessResponse) {
+                callback(response)
             }
-        }
+        })
     }
 
-    private fun startWebsocketListener() {
-        val newWebsocketClient = WebSocketClient()
-        newWebsocketClient.start()
-        val request = ClientUpgradeRequest()
-        val socket = WebsocketHandler(mapper, roundCallbacks)
-
-        assert(websocketBaseUrl.endsWith('/'))
-
-        newWebsocketClient.connect(socket, URI.create(websocketBaseUrl + "v1/round/websocket"), request)
-        websocketClient = newWebsocketClient
-    }
-
-    private fun stopWebsocketListener() {
-        websocketClient?.stop()
-        websocketClient?.destroy()
-        websocketClient = null
-    }
-
-    protected fun feignBuilder(): Feign.Builder {
-        return Feign.builder()
-                .encoder(JacksonEncoder(mapper))
-                .decoder(JacksonDecoder(mapper))
-                .logger(Slf4jLogger())
-    }
-
-    @WebSocket
-    class WebsocketHandler(
-            private val mapper: ObjectMapper,
-            private val callbacks: Iterable<RoundCallback>
-    ) {
-        @OnWebSocketMessage
-        fun onMessage(message: String) {
-            val response = mapper.readValue(message, RoundResponse::class.java)
-            for (callback in callbacks) {
-                callback.callback(response)
-            }
-        }
+    /**
+     * Removes a tournament [callback].
+     */
+    fun removeTournamentCallback(callback: WebSocketCallback<SuccessResponse>) {
+        tournamentWebSocket.removeCallback(callback)
     }
 }
 
 class AuthenticatingRestWarsClient(hostname: String, port: Int, val username: String, val password: String) : RestWarsClient(hostname, port) {
-    private val client: Restwars = feignBuilder()
-            .requestInterceptor(BasicAuthRequestInterceptor(username, password))
-            .target(Restwars::class.java, httpBaseUrl)
+    private val client: Restwars = Restwars.create(httpBaseUrl, username, password, mapper)
+
+    init {
+        // Test authorization
+        client.getPlayer()
+    }
 
     fun listPlanets(): PlanetsResponse = client.listPlanets()
 
@@ -208,4 +130,16 @@ class AuthenticatingRestWarsClient(hostname: String, port: Int, val username: St
     fun listFlightsWithDestination(location: String) = client.listFlightsWithDestination(location)
 
     fun listFlightsWithStart(location: String) = client.listFlightsWithStart(location)
+
+    fun listFights(since: Long): FightsResponse = client.listFights(since)
+
+    fun listFightsWithPlanet(location: String, since: Long): FightsResponse = client.listFightsWithPlanet(location, since)
+
+    fun listDetectedFlights(): DetectedFlightsResponse = client.listDetectedFlights()
+
+    fun listDetectedFlights(since: Long): DetectedFlightsResponse = client.listDetectedFlights(since)
+
+    fun listEvents(): EventsResponse = client.listEvents()
+
+    fun listEvents(since: Long): EventsResponse = client.listEvents(since)
 }