+++ server/src/main/kotlin/restwars/rest/api/Flights.kt
@@ -18,7 +18,8 @@ fun FlightsResponse.Companion.from(flights: List<Flight>) = FlightsResponse(flig
 
 fun FlightResponse.Companion.fromFlight(flight: Flight) = FlightResponse(
         LocationResponse.fromLocation(flight.start), LocationResponse.fromLocation(flight.destination),
-        flight.arrivalInRound, ShipsResponse.fromShips(flight.ships)
+        flight.arrivalInRound, ShipsResponse.fromShips(flight.ships), flight.direction.name,
+        ResourcesResponse.fromResources(flight.cargo)
 )
 
 fun DetectedFlightsResponse.Companion.fromFlights(flights: List<DetectedFlightWithFlight>) = DetectedFlightsResponse(flights.map { DetectedFlightResponse.fromFlight(it) })