+++ server/src/test/kotlin/restwars/doc/GenerateDocs.kt
@@ -2,6 +2,7 @@ package restwars.doc
 
 import restwars.business.building.BuildingType
 import restwars.business.event.EventType
+import restwars.business.flight.FlightDirection
 import restwars.business.flight.FlightType
 import restwars.business.ship.ShipType
 import restwars.rest.api.*
@@ -70,7 +71,7 @@ object GenerateDocs {
         writeToFile(Type.RESPONSE, "start-flight.json", FlightResponse(LocationResponse(1, 2, 3), LocationResponse(4, 5, 6), 112, ShipsResponse(listOf(
                 ShipResponse(ShipType.MOSQUITO.name, 3),
                 ShipResponse(ShipType.COLONY.name, 2))
-        )))
+        ), FlightDirection.OUTWARD.name, ResourcesResponse(0, 0, 0)))
         writeToFile(Type.RESPONSE, "telescope-scan.json", ScanResponse(listOf(
                 ScannedPlanetResponse(LocationResponse(1, 1, 1), "player1"),
                 ScannedPlanetResponse(LocationResponse(1, 1, 2), "player2"),
@@ -93,19 +94,19 @@ object GenerateDocs {
                 FlightResponse(LocationResponse(1, 2, 3), LocationResponse(4, 5, 6), 112, ShipsResponse(listOf(
                         ShipResponse(ShipType.MOSQUITO.name, 3),
                         ShipResponse(ShipType.COLONY.name, 2))
-                )),
+                ), FlightDirection.OUTWARD.name, ResourcesResponse(0, 0, 0)),
                 FlightResponse(LocationResponse(1, 2, 3), LocationResponse(7, 8, 9), 144, ShipsResponse(listOf(
                         ShipResponse(ShipType.MOSQUITO.name, 1),
                         ShipResponse(ShipType.MULE.name, 1))
-                ))
+                ), FlightDirection.OUTWARD.name, ResourcesResponse(0, 0, 0))
         )))
-        writeToFile(Type.RESPONSE, "metadata-ships.json", ShipsMetadata(listOf(
-                ShipMetadata("MOSQUITO", 12, 1.0, ResourcesResponse(10, 5, 40), 1.0, 14, 1, 50),
-                ShipMetadata("MULE", 24, 0.5, ResourcesResponse(20, 10, 80), 1.2, 1, 30, 500)
+        writeToFile(Type.RESPONSE, "metadata-ships.json", ShipsMetadataResponse(listOf(
+                ShipMetadataResponse("MOSQUITO", 12, 1.0, ResourcesResponse(10, 5, 40), 1.0, 14, 1, 50),
+                ShipMetadataResponse("MULE", 24, 0.5, ResourcesResponse(20, 10, 80), 1.2, 1, 30, 500)
         )))
-        writeToFile(Type.RESPONSE, "metadata-buildings.json", BuildingsMetadata(listOf(
-                BuildingMetadata("COMMAND_CENTER", 1, 50, ResourcesResponse(100, 50, 400)),
-                BuildingMetadata("CRYSTAL_MINE", 1, 25, ResourcesResponse(50, 25, 200))
+        writeToFile(Type.RESPONSE, "metadata-buildings.json", BuildingsMetadataResponse(listOf(
+                BuildingMetadataResponse("COMMAND_CENTER", 1, 50, ResourcesResponse(100, 50, 400)),
+                BuildingMetadataResponse("CRYSTAL_MINE", 1, 25, ResourcesResponse(50, 25, 200))
         )))
         writeToFile(Type.RESPONSE, "tournament-started.json", SuccessResponse("Tournament has started"))
         writeToFile(Type.RESPONSE, "points.json", PointsResponse(listOf(
@@ -120,6 +121,7 @@ object GenerateDocs {
                 EventResponse(UUID.fromString("06319f52-9dd6-488f-bf84-2757e92217b4"), EventType.BUILDING_COMPLETE.name, 120, LocationResponse(1, 2, 3)),
                 EventResponse(UUID.fromString("09deb43b-f6fc-4f80-9651-00c766d67d97"), EventType.PLANET_COLONIZED.name, 224, LocationResponse(4, 5, 6))
         )))
+        writeToFile(Type.RESPONSE, "player.json", PlayerResponse("player1"))
     }
 }